scope-udp-prompt/pyproject.toml
[project]
name = "scope-udp-prompt"
version = "0.1.0"
description = "Receives text via UDP and injects it as a prompt for downstream pipelines"
requires-python = ">=3.12"
dependencies = ["scope-bus"]

[project.entry-points."scope"]
scope_udp_prompt = "scope_udp_prompt"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scope_udp_prompt"]
